requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
ansicolors==1.1.8
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
channels==3.0.4
channels-redis==3.4.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.3
django_heroku==0.3.1
dj_rest_auth==2.2.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-oauth==1.1.0
fcm-django==1.0.8
firebase-admin==5.2.0
frozenlist==1.3.0
google-api-core==2.7.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
msgpack==1.0.3
multidict==6.0.2
mypy==0.931
mypy-extensions==0.4.3
proto-plus==1.20.3
protobuf==3.19.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
tomli==2.0.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
web-command==0.1.0
yarl==1.7.2
